Move the extend class into parent namespace.
