+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/LanguageExtend.cs
@@ -6,9 +6,10 @@ using System.Collections.Generic;
 using osu.Framework.Allocation;
 using osu.Framework.Bindables;
 using osu.Framework.Graphics;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Language;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Language
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends
 {
     public class LanguageExtend : EditExtend
     {
