+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Extends/ReferenceExtend.cs
@@ -3,8 +3,9 @@
 
 using System.Collections.Generic;
 using osu.Framework.Graphics;
+using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Reference;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Reference
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends
 {
     public class ReferenceExtend : EditExtend
     {
