+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/LyricEditor.cs
@@ -16,13 +16,6 @@ using osu.Framework.Input.Events;
 using osu.Game.Rulesets.Karaoke.Configuration;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Language;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Notes;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Reference;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.RubyRomaji;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Singers;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.Texting;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Extends.TimeTags;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Extensions;
